+++ common/config/web.php
@@ -8,7 +8,8 @@ $config = [
         ]
     ],
     'as locale' => [
-        'class' => 'common\behaviors\LocaleBehavior'
+        'class' => 'common\behaviors\LocaleBehavior',
+        'enablePreferredLanguage' => true
     ]
 ];